Move voiceDesc to its own file
