+++ src/importexport/musicxml/internal/musicxml/musicxmlpart.h
@@ -25,6 +25,7 @@
 #include "types/fraction.h"
 #include "dom/interval.h"
 #include "musicxmltypes.h"
+#include "musicxmlvoicedesc.h"
 
 namespace mu::engraving {
 class MusicXmlOctaveShiftList : public std::map<Fraction, int>
